+++ executor/index_lookup_merge_join_test.go
@@ -1,12 +1,8 @@
 package executor_test
 
 import (
-	"strings"
-
-	"context"
 	. "github.com/pingcap/check"
 	"github.com/pingcap/failpoint"
-	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/util/testkit"
 )
 
@@ -49,58 +45,4 @@ func (s *testSuite9) TestIssue18068(c *C) {
 	// Do not hang in index merge join when the second and third execute.
 	tk.MustExec("select  /*+ inl_merge_join(s)*/ 1 from t join s on t.a = s.a limit 1")
 	tk.MustExec("select  /*+ inl_merge_join(s)*/ 1 from t join s on t.a = s.a limit 1")
-}
-
-func (s *testSuite9) TestIssue18572_1(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1(a int, b int, index idx(b));")
-	tk.MustExec("insert into t1 values(1, 1);")
-	tk.MustExec("insert into t1 select * from t1;")
-
-	c.Assert(failpoint.Enable("github.com/pingcap/tidb/executor/testIndexHashJoinInnerWorkerErr", "return"), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("github.com/pingcap/tidb/executor/testIndexHashJoinInnerWorkerErr"), IsNil)
-	}()
-
-	rs, err := tk.Exec("select /*+ inl_hash_join(t1) */ * from t1 right join t1 t2 on t1.b=t2.b;")
-	c.Assert(err, IsNil)
-	_, err = session.GetRows4Test(context.Background(), nil, rs)
-	c.Assert(strings.Contains(err.Error(), "mockIndexHashJoinInnerWorkerErr"), IsTrue)
-}
-
-func (s *testSuite9) TestIssue18572_2(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1(a int, b int, index idx(b));")
-	tk.MustExec("insert into t1 values(1, 1);")
-	tk.MustExec("insert into t1 select * from t1;")
-
-	c.Assert(failpoint.Enable("github.com/pingcap/tidb/executor/testIndexHashJoinOuterWorkerErr", "return"), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("github.com/pingcap/tidb/executor/testIndexHashJoinOuterWorkerErr"), IsNil)
-	}()
-
-	rs, err := tk.Exec("select /*+ inl_hash_join(t1) */ * from t1 right join t1 t2 on t1.b=t2.b;")
-	c.Assert(err, IsNil)
-	_, err = session.GetRows4Test(context.Background(), nil, rs)
-	c.Assert(strings.Contains(err.Error(), "mockIndexHashJoinOuterWorkerErr"), IsTrue)
-}
-
-func (s *testSuite9) TestIssue18572_3(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1(a int, b int, index idx(b));")
-	tk.MustExec("insert into t1 values(1, 1);")
-	tk.MustExec("insert into t1 select * from t1;")
-
-	c.Assert(failpoint.Enable("github.com/pingcap/tidb/executor/testIndexHashJoinBuildErr", "return"), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("github.com/pingcap/tidb/executor/testIndexHashJoinBuildErr"), IsNil)
-	}()
-
-	rs, err := tk.Exec("select /*+ inl_hash_join(t1) */ * from t1 right join t1 t2 on t1.b=t2.b;")
-	c.Assert(err, IsNil)
-	_, err = session.GetRows4Test(context.Background(), nil, rs)
-	c.Assert(strings.Contains(err.Error(), "mockIndexHashJoinBuildErr"), IsTrue)
 }
